Creation of Pyramid Type
